+++ NEWS.md
@@ -77,8 +77,8 @@
   They also gain `version` and `compare` arguments to supply requirements
   programmatically.
 
-* `check_installed()` gains a `src` argument to specificy pak-style
-  sources.
+* `check_installed()` gains an `action` argument that is called when
+  the user chooses to install and update missing and outdated packages.
 
 * `abort()`, `warn()`, and `inform()` gain a `body` argument to supply
   additional bullets in the error message.